Improvements and bugfixes
